feat: build level-2 features - add, list
Currently using an ArrayList<String> to contain the added messages. Extended on existing methods and created new methods to handle the new cases. Also abstracted the input reading to a separate function. Plus added more personality to the 'chatbot'

Issue: need to fix default syntax with the switch case to follow the guidelines proposed. Need to look into creating a special print method as well to make the replies more recognizable.