feat+refactor: refactor duchess awaitInput and incorporate DuchessException
Changed implementation of awaitInput() to be iterative instead of recursive. Added use of DuchessExceptions to handle the following cases:

No deadline /by or event /at.

No description for any task.

Completion of task that does not exist/out of range

Unrecognized input was already handled, no change to that implementation. This meets Level 5 requirements.