feat: build datetimestringformatter to provide meaningful dates
Built a helper class DateTimeStringFormatter with static method formatDateTime that returns a meaningful string from a given LocalDateTime. The standards would be as follows:

If the deadline is overdue, it'll say [OVERDUE].

If the deadline is today, it will say today.

If the deadline is yesterday, it will say yesterday.

If the deadline is tomorrow, it will say tomorrow.

If the deadline is within 1 week, it will say the day of the week e.g. Mon Tues.

If the deadline is within the same year, the year is omitted.

Else it will return the full date and time.

To do: Need to touch up the Duchess class.