perf: add exception handling for tasklist removetask and gettask
Added an additional check for the index in the getTask and removeTask operations. Although this is not supposed to happen in the Duchess program due to prior checks, this is for future extensions which may utilise the TaskList class.

This change causes Ui and CommandHandler to break. Exception handling will need to be added to their methods as well.