test: write junit test for storage
The JUnit test for Storage is slightly more complicated than the rest. It looks out for the following:
- Trying to load a save file from a folder that does not exist
- Trying to load a save file that does not exist
- Successfully loading a save file
- Trying to load a save file in a nested directory when no folders exist
- Saving to an invalid file path
- Saving to a valid file path

The cleanUp() method is run after all tests and helps to delete all folders and files created in the process of testing.

Currently lacking the ability to test the exception handling of Storage when a file path so incorrect is passed in that it causes Storage to be unable to generate the required folders. The exception that states that saving will be not possible should be thrown up.