feat: extend duke from javafx application
Previously, the project used to run from Duke.

Changes were made to Duke such that it extends from JavaFX Application, and now has the start() method. Implemented Tutorials 1 and 2 of the JavaFX tutorials, but changed the implementation of the window to follow the flexbox concept - using VBox and HBox to allow for responsiveness.

To do: Continue with tutorial to implement everything else.