Adding two methods for printing the message or just return the message in String format for all Commands and Ui.
Furthermore, add new method for the ReplyFormat to include the outline for the message or not, to cater for both the terminal and GUI.

Add two profile picture for duke and user for the GUI at image folder.

Add two fxml for DialogBox and MainWindow in view folder.

Add three new classes: DialogBox.java, Launcher.java and MainWindow.java for the GUI implementation.

Add new class, Main.java, to acts as bridge for the Duke logic and GUI.